+++ CHANGELOG.md
@@ -27,6 +27,8 @@
   have no incompatible flag combinations.
 * **\[core\]** Add the `requireEnvironment` parameter to the `@Action` annotation to control whether the
   `Flags.ENVIRONMENT_REQUIRED` flag is added when invoking the action.
+* **\[core\]** Add the `requireEnvironment` parameter to the `@Computed` annotation to control whether the
+  `Flags.ENVIRONMENT_REQUIRED` flag is added when computing the value.
 
 ### [v0.108](https://github.com/arez/arez/tree/v0.108) (2018-09-27)
 [Full Changelog](https://github.com/arez/arez/compare/v0.107...v0.108)